Generics
git-svn-id: svn://svn.liferay.com/repos/public/portal/trunk@14403 05bdf26c-840f-0410-9ced-eb539d925f36

+++ portal-impl/src/com/liferay/portlet/imagegallery/service/persistence/IGImageFinderImpl.java
@@ -87,7 +87,7 @@ public class IGImageFinderImpl implements IGImageFinder {
 			QueryPos qPos = QueryPos.getInstance(q);
 
 			for (int i = 0; i < folderIds.size(); i++) {
-				Long folderId = (Long)folderIds.get(i);
+				Long folderId = folderIds.get(i);
 
 				qPos.add(folderId);
 			}
